+++ lib/screens/search_clothes_screen.dart
@@ -53,7 +53,7 @@ class _SearchClothesScreenState extends State<SearchClothesScreen> {
     final list = await _firebaseService.getUserCustomCategories();
     if (mounted) {
       setState(() {
-        _customCategories = list;
+        _customCategories = List<String>.from(list);
       });
     }
   }